Converted engineCheckmatePractice.js to typescript (not tested for crashes)
